♻️ refactor(tui): move the Exec picker into views/exec_picker.rs
Second of the two pilot views (#635), and the one that repatriates
satellite fields. `state/exec_picker.rs`, `draw_exec_picker` from `ui.rs`
and the four `App` methods from `app.rs` become one module.

`exec_picker_cfg`, `exec_picker_common_dir` and `exec_container_seq` were
satellites of a state type that had already been extracted — the shape the
issue names as most of what makes `App` 85 fields wide. They are now
fields of `ExecPicker`, reached directly from the `impl App` block since
it lives in the same module as the struct, so no accessors were added.
`App` goes from 85 fields to 82 (the issue says 83; it was measured at 85
on `dev` at 0409be6c).

The capture is a named `ExecPicker::capture_context` rather than two
assignments through `open`, because it is a different question: `open`
re-populates the list and rewinds the highlight on every open, while the
capture pins the `[exec]` config and the commondir an auto-refresh or a
workspace repo swap must not drift under the overlay (Codex #333 review).

`container_seq` is the one field `open` must NOT touch: it is the
monotonic half of a run's container name (#421), and the pid alone would
collide across two overlays opened on the same worktree in one session.
It was a separate `App` field before, so `open` never reached it; now that
it sits next to `selected`, that is an invariant someone can break by
accident. The next commit pins it.

`picker_lines` and `push_modal_hint` widen to `pub(crate)` for the moved
renderer. `views` is a private module, so `gwm::tui::ExecPicker` and every
other public path is unchanged.

No behaviour change. `cargo test`: 111 suites, 3530 passed, 0 failed.
`cargo clippy --all-targets -- -D warnings` and `cargo fmt --check`: clean.

refs #635
