+++ src/tui/mod.rs
@@ -51,7 +51,6 @@ pub use state::config_panel::{
 };
 pub use state::confirm::{ConfirmButton, ConfirmKeyAction, ConfirmModal, CountdownTickOutcome};
 pub use state::create_form::{CreateForm, Field};
-pub use state::exec_picker::ExecPicker;
 pub use state::filter::FilterState;
 pub use state::github_fetch::{FetchKey, GitHubFetch, GitHubFetchState};
 pub use state::link_prompt::LinkPrompt;
@@ -63,6 +62,7 @@ pub use views::commits::{
   commit_meta_columns, commits_meta_pick, recent_commits_lines, recent_commits_listing, CommitsModal, CommitsSnapshot,
   COMMITS_MAX, COMMITS_PAGE,
 };
+pub use views::exec_picker::ExecPicker;
 
 /// Ordered list of clipboard tools to try for the host OS (issue #73).
 /// First entry that resolves on `$PATH` wins. Returned in the
