+++ src/tui/state/mod.rs
@@ -23,7 +23,6 @@ pub mod config_panel;
 pub mod confirm;
 pub mod create_form;
 pub mod detail_overlay;
-pub mod exec_picker;
 pub mod filter;
 pub mod github_fetch;
 pub mod link_prompt;
